엔티티 수정

+++ src/main/java/com/wanted/preonboarding/domain/dto/reservation/NotificationResponseDto.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 @Getter
 public class NotificationResponseDto {
   private UserDto user;
-  private long    notificationId;
+  private Integer notificationId;
 }
